stdlib continue

+++ xlibc/include/stddef.h
@@ -5,8 +5,8 @@
 #define offsetof(st, m) \
     ((size_t)&(((st *)0)->m))
 
-typedef unsigned long long size_t;
-typedef long long ssize_t;
+typedef unsigned long size_t;
+typedef long ssize_t;
 
 typedef long ptrdiff_t;
 typedef int wchar_t;
